Train a CNN on a directory of images This function trains a keras CNN model on a folder of n-dimensional images and returns the trained model together with validation data

Train a CNN on a directory of images This function trains a keras CNN model on a folder of n-dimensional images and returns the trained model together with validation data

Usage

train_cnn(
  model,
  image_directory = NULL,
  image_list = NULL,
  validation_split = 0.2,
  epochs = 10,
  weight_classes = FALSE,
  sub_cropsize = NULL,
  channels = 1:5,
  batch_size = 128,
  scale = NULL,
  log_scale = FALSE,
  featurewise_standardisation = FALSE,
  imagewise_normalisation = FALSE,
  subsample_training_data = NULL,
  early_stopping = TRUE,
  ...
)

Arguments

model

A keras CNN model

image_directory

The directory containing folders of images named by class

image_list

Alternatively, a list of pre-loaded crops may be supplied

validation_split

The proportion of images that should be saved for validation

epochs

Training epochs

weight_classes

Should the loss function be weighted to account for unbalanced classes

sub_cropsize

If the images should be further cropped, specify the number of pixels

channels

Specify the names of the channels that should be read into the model

batch_size

Batch size

scale

Manually specify number to divide pixel values by

log_scale

Log scale intensity values

featurewise_standardisation

Argument passed to image generator

imagewise_normalisation

Argument passed to image generator

subsample_training_data

Argument passed to image generator

early_stopping

Stop the training early once the loss function stabilises

...

Other image normalsiation arguments

Value

Returns a list containing a trained keras model, training history, confusion matrix (from validation images), validation image predictions and names